Replace template functor with std::function<void()>
